expansions step9 done - only user that created a blog, can delete it

+++ controllers/blogs.js
@@ -51,11 +51,28 @@ blogsRoute.post("/", async (request, response, next) => {
 });
 
 blogsRoute.delete("/:id", async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-  } catch (error) {
-    next(error);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({
+      error: "Token is missing or invalid",
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === user._id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } catch (error) {
+      next(error);
+    }
+  } else {
+    response.status(401).json({
+      error: "Unauthorized access",
+    });
   }
 });
